Release of 1.9.2.
Bugfixes
- Bug when useing XAMPP as the server
-- Removed use of old mime-content-type(string) method because its
deprecated.
- Bug fix when you tried to copy a file containing a sub sub [...]
directory

Feature additions
- Copy of directories now supported
- Mainteance mode added
- Guest account support added
- New configuration values added
- Moved any possible settings value out of Program.inc.php
- Removed the title "cloud storage" out of the logo, because Redundancy
is not a service, its a program ;).
- Added a small manual how to install redundancy on a webserver. A
bigger manual (with pictures) will follow.

The rating of this build is still "unstable". Use with caution!

+++ Includes/share.inc.php
@@ -7,13 +7,13 @@
 	if (isset($_SESSION["user_id"]))
 		$userID = $_SESSION["user_id"];	
 	$code = "";
-	if (isset($_GET["delete"]) && $_GET["delete"] == "true")
+	if (isset($_GET["delete"]) && $_GET["delete"] == "true" && $_SESSION["role"] != 3)
 	{					
 		mysql_query("delete from `Share` where  Hash = '$file' and UserID = '$userID'") or die("Error: 026");
 		mysql_close($connect);	
 		header ("Location: index.php?module=list");
 	}
-	else if (isset($_GET["new"]) && $_GET["new"] == "true")
+	else if (isset($_GET["new"]) && $_GET["new"] == "true" && $_SESSION["role"] != 3)
 	{
 		mysql_query("Select *  from `Share` where  Hash = '$file' and UserID = '$userID'")  or die("Error: 027");
 		if (mysql_affected_rows() > 0)
@@ -46,7 +46,7 @@
 	{	
 		$result = mysql_query("Select * from Share  where UserID = '$userID' and Hash ='".$file."'") or die("Error: 029".mysql_error());	
 		while ($row = mysql_fetch_object($result)) {
-			$sharetext = $_SERVER["SERVER_NAME"].$_GLOBALS["Program_Storage"]."index.php?share=".$row->Extern_ID;
+			$sharetext = $_SERVER["SERVER_NAME"].$_SESSION["config"]["Program_Share_Dir"]."index.php?share=".$row->Extern_ID;
 			echo "<textarea cols='70' rows='2'>$sharetext</textarea><br>";
 			echo "<a href = 'index.php?module=file&file=".$row->Hash."'>View</a> ";			
 			echo "<a href = 'index.php?module=share&file=".$row->Hash."&delete=true'>Unshare</a>";			
@@ -81,4 +81,5 @@
 		} 
 		mysql_close($connect);			
 	}
+	header("Location: index.php?module=list");
 ?>
